moved resposne and request to modules. Create a Request class which holds turns request params into object attributes recursively
